Fixed color order
The order in which the colors are getting set seems to play a role.

From :h cterm-colors:
<	When setting the "ctermbg" color for the Normal group, the
	'background' option will be adjusted automatically.  This causes the
	highlight groups that depend on 'background' to change!  This means
	you should set the colors for Normal first, before setting other
	colors.

In my case I had base16-tomorrow installed with airline and the whitespace warning of airline was hard to read. It turned out that this was the root cause for that: I had :set background=light before :colorscheme base16-tomorrow in my vimrc and :verbose set background? revealed, that base16-tomorrow.vim had reset it back to dark, even though it does not contain any set background-command.